requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==3.2.9
django-crispy-forms==1.13.0
django-tinymce4-lite==1.8.0
jsmin==3.0.0
Pillow==8.4.0
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
